pyproject.toml
[project]
name = "food-recs"
version = "0.1.0"
description = "Item-to-item food recommendation system based on basket analysis"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "gensim>=4.3.0",
    "tqdm>=4.65.0",
    "streamlit>=1.28.0",
    "plotly>=5.18.0",
    "scikit-learn>=1.3.0",
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "fire>=0.5.0",
    "mlflow>=2.9.0",
    "kaleido>=1.0.0",
    "dvc[s3]>=3.64.0",
    "implicit>=0.7.2",
    "scipy>=1.11.0",
    "sentence-transformers>=2.2.2",
    "lightgbm>=3.3.2",
    "torch>=1.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
]

[project.scripts]
food-recs = "food_recs.commands:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["food_recs"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
